Cleaned-up tree traversal function
Code blocks inherit their script lexical scope, but also have access to the scope of the script or function calling them.

I had to keep `_q` inside the traversal function because it would be shadowed by the outer `_queue` defined below in the `_nodeCode` Code call, if the variables had the same name.